[#192] refactor: resolver, converter, interceptor 리페키징
