+++ src/main/java/com/sparta/popupstore/domain/common/converter/PopupStoreSearchTypeConverter.java
@@ -1,4 +1,4 @@
-package com.sparta.popupstore.domain.popupstore.controller.converter;
+package com.sparta.popupstore.domain.common.converter;
 
 import com.sparta.popupstore.domain.common.exception.CustomApiException;
 import com.sparta.popupstore.domain.common.exception.ErrorCode;
