+++ src/main/java/com/sparta/popupstore/domain/common/resolver/AuthCompanyResolver.java
@@ -1,12 +1,10 @@
-package com.sparta.popupstore.config;
+package com.sparta.popupstore.domain.common.resolver;
 
-import com.sparta.popupstore.domain.common.annotation.AuthSocialUser;
+import com.sparta.popupstore.domain.common.annotation.AuthCompany;
 import com.sparta.popupstore.domain.common.exception.CustomApiException;
 import com.sparta.popupstore.domain.common.exception.ErrorCode;
-import com.sparta.popupstore.domain.oauth2.controller.converter.OAuth2PlatformConverter;
-import com.sparta.popupstore.domain.oauth2.entity.SocialUser;
-import com.sparta.popupstore.domain.oauth2.repository.SocialUserRepository;
-import com.sparta.popupstore.domain.oauth2.type.OAuth2Platform;
+import com.sparta.popupstore.domain.company.entity.Company;
+import com.sparta.popupstore.domain.company.repository.CompanyRepository;
 import com.sparta.popupstore.jwt.JwtUtil;
 import io.jsonwebtoken.Claims;
 import lombok.NonNull;
@@ -20,16 +18,15 @@ import org.springframework.web.method.support.ModelAndViewContainer;
 
 @Component
 @RequiredArgsConstructor
-public class AuthSocialUserResolver implements HandlerMethodArgumentResolver {
+public class AuthCompanyResolver implements HandlerMethodArgumentResolver {
 
-    private final SocialUserRepository socialUserRepository;
-    private final OAuth2PlatformConverter oAuth2PlatformConverter;
+    private final CompanyRepository companyRepository;
     private final JwtUtil jwtUtil;
 
     @Override
     public boolean supportsParameter(MethodParameter parameter) {
-        return parameter.hasParameterAnnotation(AuthSocialUser.class)
-                && parameter.getParameterType().equals(SocialUser.class);
+        return parameter.hasParameterAnnotation(AuthCompany.class)
+                && parameter.getParameterType().equals(Company.class);
     }
 
     @Override
@@ -39,13 +36,9 @@ public class AuthSocialUserResolver implements HandlerMethodArgumentResolver {
             @NonNull NativeWebRequest webRequest,
             WebDataBinderFactory binderFactory
     ) {
-        Claims socialUserInfo = jwtUtil.getInfoFromWebRequest(webRequest);
+        Claims companyInfo = jwtUtil.getInfoFromWebRequest(webRequest);
 
-        String providerId = socialUserInfo.getSubject();
-        OAuth2Platform platform = oAuth2PlatformConverter.convert(
-                (String) socialUserInfo.get(JwtUtil.OAUTH2_PLATFORM_KEY)
-        );
-        return socialUserRepository.findByPlatformAndPlatformId(platform, providerId)
+        return companyRepository.findByEmailAndDeletedAtIsNull(companyInfo.getSubject())
                 .orElseThrow(() -> new CustomApiException(ErrorCode.NEED_LOGIN));
     }
 }
